+++ drivers/ethernet/CMakeLists.txt
@@ -53,6 +53,8 @@ endif()
 
 if(CONFIG_ETH_STM32_HAL)
   zephyr_library_sources(eth_stm32_hal_common.c)
+  zephyr_library_sources_ifdef(CONFIG_ETH_STM32_HAL_API_V1	eth_stm32_hal_v1.c)
+  zephyr_library_sources_ifdef(CONFIG_ETH_STM32_HAL_API_V2	eth_stm32_hal_v2.c)
   zephyr_library_sources_ifdef(CONFIG_PTP_CLOCK_STM32_HAL	eth_stm32_hal_ptp.c)
 endif()
